+++ spec-integration/common.ts
@@ -1,0 +1,8 @@
+import { existsSync } from 'fs'
+import { config } from 'dotenv';
+
+export const setEnvs = () => {
+  if (existsSync('.env')) {
+    config();
+  }
+}