+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.3.0 (December 17, 2021)
+* Now method `uploadAttachment` from `AttachmentProcessor` save attachments directly to `Maester`
+* Now method `uploadAttachment` from `AttachmentProcessor` accepts additional argument `contentType`
+
 ## 1.2.0 (November 12, 2021)
 * Added `fetchAllSecretsForWorkspace` method to `PlatformApiLogicClient`